Added README and JSON output scripts
- README added to root directory

- create_actors_json.py added: Goes through all actor information in the
database and outputs the relevant information to JSON files

- create_movies_json.py added: Goes through all movie information in the
database and outputs the relevant information to JSON files

- create_random_movies_json.py added: Splits the movie ID range of the
database into disjunct blocks, then selects random sub-samples of each
block and outputs the relevant movie information to JSON files. This is
not exhaustive, but is considerably faster for gathering at least a
fraction of the entire database's movie data.

- settings.py added: Contains some constants commonly used.

- utils.py added: Helper functions